feat(sponsor): filter table by currency
