+++ app/sponsor/[sponsorId]/components/sponsor-history-cards/sponsor-history-cards.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 
 import { SponsorHistoryCard } from "app/sponsor/[sponsorId]/components/sponsor-history-card/sponsor-history-card";
-import { useSponsorHistory } from "app/sponsor/[sponsorId]/hooks/use-sponsor-history";
+import { useSponsorHistory } from "app/sponsor/[sponsorId]/hooks/use-sponsor-history/use-sponsor-history";
 
 import { IMAGES } from "src/assets/img";
 import { ShowMore } from "src/components/Table/ShowMore";
