+++ app/sponsor/[sponsorId]/features/sponsor-section-budget/sponsor-section-budget.tsx
@@ -1,27 +1,12 @@
-import { isString } from "lodash";
-import { useParams } from "next/navigation";
-
 import { SponsorBudgetCard } from "app/sponsor/[sponsorId]/components/sponsor-budget-card/sponsor-budget-card";
 import { SponsorDepositCard } from "app/sponsor/[sponsorId]/components/sponsor-deposit-card/sponsor-deposit-card";
-
-import SponsorApi from "src/api/Sponsors";
+import { useSponsorDetail } from "app/sponsor/[sponsorId]/hooks/use-sponsor-detail/use-sponsor-detail";
 
 import { Translate } from "components/layout/translate/translate";
 import { Typography } from "components/layout/typography/typography";
 
 export function SponsorSectionBudget() {
-  const { sponsorId } = useParams();
-
-  const sponsorIdIsString = isString(sponsorId);
-
-  const { data, isLoading } = SponsorApi.queries.useGetSponsorById({
-    params: {
-      sponsorId: sponsorIdIsString ? sponsorId : "",
-    },
-    options: {
-      enabled: sponsorIdIsString,
-    },
-  });
+  const { data, isLoading } = useSponsorDetail();
 
   function renderBudgets() {
     if (isLoading) {
