+++ app/sponsor/[sponsorId]/hooks/use-sponsor-detail/use-sponsor-detail.ts
@@ -1,20 +1,16 @@
 import { isString } from "lodash";
 import { useParams } from "next/navigation";
 
-import { TUseSponsorHistory } from "app/sponsor/[sponsorId]/hooks/use-sponsor-history.types";
-
 import SponsorApi from "src/api/Sponsors";
 
-export function useSponsorHistory(props?: TUseSponsorHistory.Props) {
-  const { queryParams } = props ?? {};
-
+export function useSponsorDetail() {
   const { sponsorId } = useParams();
+
   const sponsorIdIsString = isString(sponsorId);
 
-  return SponsorApi.queries.useGetSponsorTransactions({
+  return SponsorApi.queries.useGetSponsorById({
     params: {
       sponsorId: sponsorIdIsString ? sponsorId : "",
-      queryParams,
     },
     options: {
       enabled: sponsorIdIsString,
